pyproject.toml
[project]
name = "jira-extractor"
version = "0.1.0"
description = "Extract data from Jira Server API and load into CSV or PostgreSQL"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "pydantic-settings>=2.3.0",
    "click>=8.1.0",
    "python-json-logger>=2.0.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
scheduler = [
    "apscheduler>=3.10.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
jira-extractor = "jira_extractor.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jira_extractor"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
